Fix components
