+++ docs/app/(docs)/[[...slug]]/(home)/components/agents/index.tsx
@@ -98,7 +98,7 @@ const row1 = logos.slice(0, HALF);
 const row2 = logos.slice(HALF);
 
 export const Agents = () => (
-  <div className="grid gap-16">
+  <div className="grid gap-16 py-16 px-8">
     <div className="mx-auto grid max-w-3xl gap-4 text-center">
       <h2 className="font-semibold text-4xl tracking-tight">
         Works with all your favourite agents
