+++ docs/app/(docs)/[[...slug]]/(home)/components/footer.tsx
@@ -1,5 +1,5 @@
 export const Footer = () => (
-  <footer className="text-center">
+  <footer className="text-center p-8">
     <p className="text-muted-foreground text-sm">
       Made with ❤️ and ☕ by{" "}
       <a
